+++ tools/ipod-boot/flsh.sh
@@ -17,7 +17,7 @@ RES="$ROOT/resources"
 
 : "${TRACE:=$HOME/dev/.cargo-target/release/trace}"
 : "${IMG:=diag}"
-: "${FLASH:=$RES/internal_rom_000000-0FFFFF/internal_rom_000000-0FFFFF.bin}"
+: "${FLASH:=$RES/roms/retail_5g_MA146_HwVr000B0005_internal_rom_000000-0FFFFF.bin}"
 : "${DISK:=$RES/drives/ipod8g.img}"
 : "${BUDGET:=200000000}"